--- bootstrapper/internal/kubernetes/k8sapi/kubectl/generator.go
@@ -1,17 +1,0 @@
-/*
-Copyright (c) Edgeless Systems GmbH
-
-SPDX-License-Identifier: AGPL-3.0-only
-*/
-
-package kubectl
-
-import "github.com/edgelesssys/constellation/v2/bootstrapper/internal/kubernetes/k8sapi/kubectl/client"
-
-// generator implements clientGenerator interface.
-type generator struct{}
-
-// NewClients generates a new client implementing the Client interface.
-func (generator) NewClient(kubeconfig []byte) (Client, error) {
-	return client.New(kubeconfig)
-}
