+++ bootstrapper/internal/kubernetes/k8sapi/kubectl/restclient_test.go
@@ -4,7 +4,7 @@ Copyright (c) Edgeless Systems GmbH
 SPDX-License-Identifier: AGPL-3.0-only
 */
 
-package client
+package kubectl
 
 import (
 	"errors"
